separate actions2code to its own header
